% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_examples.R
\name{foxplots_example}
\alias{foxplots_example}
\title{Load one built-in example dataset}
\usage{
foxplots_example(key)
}
\arguments{
\item{key}{One of `foxplots_examples()$key`.}
}
\value{
A data frame.
}
\description{
Load one built-in example dataset
}
\examples{
str(foxplots_example("toothgrowth"))
}
