% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{info_tip}
\alias{info_tip}
\title{A small grey "?" tooltip icon for inline help.}
\usage{
info_tip(..., placement = "right")
}
\arguments{
\item{...}{Tooltip content (passed to bslib::tooltip).}

\item{placement}{Tooltip placement ("right", "top", ...).}
}
\value{
a bslib tooltip tag.
}
\description{
A small grey "?" tooltip icon for inline help.
}
